* 把job的句柄检测机制应用出去，给get icon url用。因为这个要用到异步获取网络。所以抽象出了个jobhead * 某些代理情况下，在head回调还没结束，会突然跳到write回调，而且url还会变化。群里有例子可以测试。针对这种情况，mb的网络层做了调整。

+++ net/ActivatingObjCheck.cpp
@@ -57,7 +57,7 @@ bool ActivatingObjCheck::isActivating(intptr_t loader)
 
 int ActivatingObjCheck::genId()
 {
-    InterlockedIncrement((unsigned int *)&m_newestId);
+    InterlockedIncrement((long *)&m_newestId);
     return m_newestId;
 }
